--- a/src/centrality/closeness.jl
+++ b/src/centrality/closeness.jl
@@ -1,10 +1,10 @@
 # @doc doc"""
-#         closeness_centrality(g::AbstractGraph; normalize=true)
+#         closeness_centrality(g::AbstractGeneralGraph; normalize=true)
 #
 #     Computes closeness centrality of a graph, based on all vertices.
 #     """ ->
 function closeness_centrality(
-    g::AbstractGraph;
+    g::AbstractGeneralGraph;
     normalize=true)
 
     n_v = nv(g)