--- a/src/centrality/katz.jl
+++ b/src/centrality/katz.jl
@@ -21,7 +21,7 @@
 # TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
 # OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-function katz_centrality(g::AbstractGraph, α::Real = 0.3)
+function katz_centrality(g::AbstractGeneralGraph, α::Real = 0.3)
     nvg = nv(g)
     v = ones(Float64, nvg)
     spI = speye(Float64, nvg)