--- a/src/maxadjvisit.jl
+++ b/src/maxadjvisit.jl
@@ -16,7 +16,7 @@ end
 abstract AbstractMASVisitor <: AbstractGraphVisitor
 
 function maximum_adjacency_visit_impl!{T}(
-  graph::AbstractGraph,	                      # the graph
+  graph::AbstractGeneralGraph,	                      # the graph
   pq::Collections.PriorityQueue{Int, T},               # priority queue
   visitor::AbstractMASVisitor,                      # the visitor
   colormap::Vector{Int})                            # traversal status
@@ -41,7 +41,7 @@ function maximum_adjacency_visit_impl!{T}(
 end
 
 function traverse_graph(
-  graph::AbstractGraph,
+  graph::AbstractGeneralGraph,
   T::DataType,
   alg::MaximumAdjacency,
   s::Int,
@@ -84,7 +84,7 @@ end
 #################################################
 
 type MinCutVisitor{T} <: AbstractMASVisitor
-  graph::AbstractGraph
+  graph::AbstractGeneralGraph
   parities::AbstractArray{Bool,1}
   colormap::Vector{Int}
   bestweight::T
@@ -94,7 +94,7 @@ type MinCutVisitor{T} <: AbstractMASVisitor
   vertices::Vector{Int}
 end
 
-function MinCutVisitor{T}(graph::AbstractGraph, distmx::AbstractArray{T, 2})
+function MinCutVisitor{T}(graph::AbstractGeneralGraph, distmx::AbstractArray{T, 2})
   n = nv(graph)
   parities = falses(n)
   MinCutVisitor(graph, parities, zeros(Int,n), typemax(T), zero(T), zero(Int), distmx, Int[])
@@ -166,7 +166,7 @@ end
 #################################################
 
 # function mincut{T}(
-#   graph::AbstractGraph,
+#   graph::AbstractGeneralGraph,
 #   distmx::AbstractArray{T, 2},
 #   visitor = MinCutVisitor(graph, distmx),
 #   colormap = zeros(Int, nv(graph)))
@@ -177,7 +177,7 @@ end
 # end
 
 function mincut{T}(
-  graph::AbstractGraph,
+  graph::AbstractGeneralGraph,
   distmx::AbstractArray{T, 2})
 
   visitor = MinCutVisitor(graph, distmx)
@@ -188,18 +188,18 @@ function mincut{T}(
   return( visitor.parities, visitor.bestweight)
 end
 
-function mincut(graph::AbstractGraph)
+function mincut(graph::AbstractGeneralGraph)
   nvg = nv(graph)
   mincut(graph,DefaultDistance())
 end
 
-function maximum_adjacency_visit{T}(graph::AbstractGraph, distmx::AbstractArray{T, 2}; log::Bool=false, io::IO=STDOUT)
+function maximum_adjacency_visit{T}(graph::AbstractGeneralGraph, distmx::AbstractArray{T, 2}; log::Bool=false, io::IO=STDOUT)
   visitor = MASVisitor(io, Int[],distmx,log)
   traverse_graph(graph, T, MaximumAdjacency(), first(vertices(graph)), visitor, zeros(Int, nv(graph)))
   visitor.vertices
 end
 
-function maximum_adjacency_visit(graph::AbstractGraph; log::Bool=false, io::IO=STDOUT)
+function maximum_adjacency_visit(graph::AbstractGeneralGraph; log::Bool=false, io::IO=STDOUT)
   n = nv(graph)
   distmx = sparse(zeros(n,n))
   maximum_adjacency_visit(graph,distmx; log=log, io=io)