This PR:

* Makes a new test that implements a new graph type that does not explicitly contain the edges and vertices
* Makes modifications to core.jl to make methods as general as possible. Note that these default methods might be horribly inefficient for a new abstract type, but that can be addressed on a per-method basis by the new type - they will at least work by default.
* Tightens up some of the methods in core.jl that were relying on internal fields
* Removes some methods from graphs, digraphs.jl that could be used for general graphs if directedness was checked.

I've just been going through the functions on the "basic measures" part of the docs, will get to more complicated measures before its done.

Note that a second step after this would be to consider a type hierachy that seperates abstract undirected and abstract directed graphs, like

```julia
abstract AbstractDiGraph
abstract AbstractGraph <: AbstractDiGraph
```
or
```julia
abstract AbstractGeneralGraph
abstract AbstractDiGraph <: AbstractGeneralGraph
abstract AbstractGraph <: AbstractGeneralGraph
```

which would eliminate the `is_directed` ifs that this introduces.