@IainNZ 2015-06-29
Test failure on 0.4 seems unrelated.