@jpfairbanks 2015-06-29
I like the idea and the GridGraph is a good example of the benefits of the AbstractGraph type. The interface you have should be enough to do most graph programing. It is also simpler than a fully parameterized graph type