@IainNZ 2015-06-29
Yeah I'm halfway through some of the traversal functions and they all just work. I don't think it'll take me long to finish up and then will be ready to merge.