@yeesian 2015-06-29
:+1:

Did you mean

```julia
abstract AbstractGraph
abstract AbstractDiGraph <: AbstractGraph
abstract AbstractUndirectedGraph <: AbstractGraph  # maybe in need of a better name
```

instead of vice-versa? But yeah, I'm definitely happy with the changes here.