@IainNZ 2015-06-29
No I meant what I said, and I think what you have there is my second one with different names. The first one I propose is a more controversial choice, but I think it has some logical merit.