@yeesian 2015-06-29
> The first one I propose is a more controversial choice, but I think it has some logical merit.

I apologise for being dense, but care to elaborate abit?

Here's yet another proposal:
```
abstract EdgeType

immutable Directed <: EdgeType end
immutable Undirected <: EdgeType end

abstract AbstractGraph{T <: EdgeType}
```