@IainNZ 2015-06-29
My first proposal is simply that undirected graphs are a special case of directed graphs (one where all edges and their reverse are in the graph).