@sbromberger 2015-06-29
re: the travis failures: I'm not sure we even need DataStructures anymore. Let me test something.