@sbromberger 2015-06-29
Also - this looks great. Could I bikeshed a name for a second? It seems to me that "GraphTypes" would better be called "SimpleGraphs", since that's what the union describes. Comments?