@sbromberger 2015-06-29
There's a slight performance regression in `betweenness_centrality` which is my go-to measure. It's on the order of 2 seconds on my machine:

Original code:
```
julia> @time z = betweenness_centrality(g);
  92.737 seconds      (380 M allocations: 27230 MB, 6.23% gc time)

julia> @time z = betweenness_centrality(g);
  94.471 seconds      (380 M allocations: 27230 MB, 8.05% gc time)
```

New code:
```
julia> @time z = betweenness_centrality(g);
  93.664 seconds      (380 M allocations: 27230 MB, 6.21% gc time)

julia> @time z = betweenness_centrality(g);
  96.199 seconds      (380 M allocations: 27230 MB, 7.97% gc time)
```