@IainNZ 2015-06-29
SimpleGraphs works for me, I have no love for GraphTypes.

I wonder if that performance regression is due to the `if`s, thats the only change that could affect perf.