@sbromberger 2015-06-29
I'm not TOO worried about the performance regression, but I'm wondering whether having specialization on things like `neighbors` is better than checking directedness within the function itself.