@sbromberger 2015-06-29
@IainNZ I submitted a PR to your `ird/abstract` branch.