@sbromberger 2015-06-29
PS: re https://github.com/JuliaGraphs/LightGraphs.jl/pull/72#issue-91664280 and https://github.com/JuliaGraphs/LightGraphs.jl/pull/72#issuecomment-116379688 - 

```
abstract AbstractGeneralGraph
abstract AbstractDiGraph <: AbstractGeneralGraph
abstract AbstractGraph <: AbstractGeneralGraph
```

is my preferred approach. I did lots of timing tests at the very beginning looking at the impact of directedness on various functions, and found that keeping `DiGraph` separate from `Graph` and using specialization resulted in better performance across the board (as a related aside, my guess is that this lack of specialization is what's causing the minor perf regressions in the PR).