@IainNZ 2015-06-29
Lets go with that type hierarchy, that will allows us to remove the ifs