@sbromberger 2015-07-01
Ok, after some great off-list discussion, I think what I'd like to do is the following:

```
abstract AbstractGraph
Graph <: AbstractGraph
DiGraph <: AbstractGraph

typealias SimpleGraph Union(Graph,DiGraph)
```
and not worry about the other abstractions. I am not convinced (yet?) that LightGraphs needs to support arbitrary graph types. It was created to provide a easy, fast way of performing operations on simple graphs. At this point I'm inclined to just say, "if you need a more complex graph structure, Graphs.jl is probably for you". To excerpt from an email I sent earlier, my thoughts on why arbitrary graph structures in LightGraphs is a bad idea right now:

1. People will use it without understanding the implications and get frustrated because it’s so slow (as I did with Graphs.jl - admittedly, I was using it without understanding it, and that’s likely what caused the problems);
2. I really don’t want to have to troubleshoot someone else’s custom code;
3. Any modifications we make will have to take into account the fact that custom graph types might behave very differently, and either not implement that functionality or do something that we wouldn’t normally do;
4. People will add functionality for their particular custom type of graph, and that won’t really benefit the core of LightGraphs (which are simple graphs as defined in the README) since they won’t be defined for those graphs, and in any case trying to understand them so we can merge PRs will lead us back to issue 2;
5. Nobody seems to need it right now; and if they do,
6. Graphs.jl is designed specifically for that sort of work.

So I'll keep this open for discussion until `bfsdfs` gets merged and then will create a new PR to make these changes.