@IainNZ 2015-07-01
@sbromberger actually the conclusion (to me) was
```
type Graph
  ...
end
type DiGraph
  ...
end

typealias SimpleGraph Union(Graph,DiGraph)
```
Because the decision was to specifically not support AbstractGraphs