@sbromberger 2015-07-01
I'm ok with that as well. Is there a difference, or is it just so that folks like me don't abuse `AbstractGraph` any more?