@IainNZ 2015-07-01
No performance difference, but keeping AbstractGraph implies that there is a use for it, so if there isn't going to be a use for it then I don't think it should be there.