@jpfairbanks 2015-07-02
Was I correct in understanding that this Abstract graph was an interface based abstraction so any type that defined a few core methods could inherit from AbstractGraph and thus be used. In contrast Graphs.jl is a generics based abstraction so your type had to fit into the construction provided?