@IainNZ 2015-07-06
I'll open a new PR with the simple fix.