@sbromberger 2015-07-06
Oh, I did it already. See #79.