@sbromberger 2015-07-06
...but I may have been a bit too aggressive since I renamed things like `AbstractGraphVisitor`. Let's use yours.